Move Benchmark print to encoderStats::print()
